+++ src/apprun.ts
@@ -9,6 +9,7 @@ export interface IApp {
   start<T>(element?: Element | string, model?: T, view?: View<T>, update?: Update<T>,
     options?: { history?, rendered?: (state: T) => void}): Component<T>;
   on(name: string, fn: (...args: any[]) => void, options?: any): void;
+  off(name: string, fn: (...args: any[]) => void): void;
   run(name: string, ...args: any[]): void;
   createElement(tag: string | Function, props, ...children): VNode | VNode[];
   render(element: HTMLElement, node: VNode): void;
@@ -28,23 +29,22 @@ app.start = <T>(element?: HTMLElement | string, model?: T,  view?: View<T>, upda
     return component;
 };
 
-let _app: IApp = app;
-declare var global;
-const root = global || window;
-if (root['app'] && root['app']['start']) {
-  _app = root['app'];
-} else {
-  root['app'] = _app;
+if (!app['route']) {
+  app['route'] = route;
+  app.on('//', _ => { });
+  app.on('#', _ => { });
+  app.on('route', url => route(url));
   if (typeof document === 'object') {
     document.addEventListener("DOMContentLoaded", () => {
-      window.onpopstate = () => route(location.hash || location.pathname);
+      window.onpopstate = () => route(location.hash);
       route(location.hash);
     });
   }
 }
+
 export type StatelessComponent<T={}> = (args: T) => string | VNode | void;
 export { Component, View, Action, Update, on, update };
 export { update as event };
-export default _app;
+export default app as IApp;
 
 app.on('debug', _ => 0);