+++ src/router.ts
@@ -2,10 +2,6 @@ import app from './app';
 
 const ROUTER_EVENT = '//';
 
-app.on('//', _ => { });
-app.on('#', _ => { });
-app.on('route', url => route(url));
-
 export default function route(url: string) {
   if (!url) url = '#';
   if (url.startsWith('#')) {