+++ src/vdom-html.ts
@@ -1,16 +1,16 @@
 import { createElement, updateElement, Fragment } from './vdom-my';
 import morph = require('morphdom')
 
-export function render(element, html) {
-  console.assert(!!element);
+export function render(element, html, parent?) {
   if (typeof html === 'string') {
+    html = html.trim();
     if (element.firstChild) {
       morph(element.firstChild, html);
     } else {
       element.innerHTML = html;
     }
   } else {
-    updateElement(element, html);
+    updateElement(element, html, parent);
   }
 }
 export { createElement, Fragment };