+++ tests/router.spec.ts
@@ -56,4 +56,14 @@ describe('router', () => {
     expect(fn2).toHaveBeenCalledWith('/x', 'y', 'z');
   });
 
+  it('should not convert / to #', () => {
+    const fn1 = jasmine.createSpy('fn1');
+    const fn2 = jasmine.createSpy('fn2');
+    app.on('#', fn1);
+    app.on('/', fn2);
+    app.run('route', '/');
+    expect(fn1).not.toHaveBeenCalled();
+    expect(fn2).toHaveBeenCalled();
+  });
+
 });