+++ CHANGELOG.md
@@ -1,8 +1,23 @@
 # Change Log
 
-## 2.0 beta
+## Releases
+* apprun@latet: 1.x, stable, es5
+* apprun@beta: 2.x, stable, es2015, web component
+* apprun@next: 3.x, development, es5
+* apprun@es6: 4.x, development, es2015, web component
+
+## 2.0 beta (ES6)
 
 * Make AppRun component web component
+
+## 1.14
+
+* New logic for creating stateful component
+* Add unmount function
+* Add app.off function
+* Server-side rendering
+* Don't render vnode of false
+* Add dev-tools (dist/apprun-dev-tools.js)
 
 ## 1.13